About 페이지에 Hero를 추가한다
